+++ NEWS.md
@@ -1,5 +1,6 @@
 # SlowQuartet v0.3.0
  - Add functions to generate ternary plots from tree similarity measures
+ - Remove single "splits" from splits objects
 
 # SlowQuartet v0.2.0
  - Add `MatchingSplits` function to calculate number of splits contradicted / unresolved.